+++ src/ffi/extra/reflection_extra.cc
@@ -74,7 +74,7 @@ void MakeObjectFromPackedArgs(ffi::PackedArgs args, Any* ret) {
       size_t arg_index = search_field(field_info->name);
       void* field_addr = reinterpret_cast<char*>(ptr.get()) + field_info->offset;
       if (arg_index < keys.size()) {
-        AnyView field_value = args[arg_index * 2 + 2];
+        AnyView field_value = args[static_cast<int>(arg_index * 2 + 2)];
         field_info->setter(field_addr, reinterpret_cast<const TVMFFIAny*>(&field_value));
         keys_found[arg_index] = true;
       } else if (field_info->flags & kTVMFFIFieldFlagBitMaskHasDefault) {
